Agentic-AI/pyproject.toml
[project]
name = "Agentic-AI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=23.0.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.104.0",
    "flake8>=6.0.0",
    "groq>=0.4.0",
    "httpx>=0.25.0",
    "isort>=5.12.0",
    "loguru>=0.7.0",
    "onnxruntime>=1.16.0",
    "pillow>=10.0.0",
    "prometheus-client>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "reportlab>=4.0.0",
    "requests>=2.31.0",
    "torch>=2.1.0",
    "ultralytics>=8.3.202",
    "uvicorn[standard]>=0.24.0",
]
